requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
flask-restx==1.1.0
mysql-connector-python==8.1.0
python-dotenv==1.0.0
openai==0.28.0
requests==2.31.0
Werkzeug==2.3.7